combine data files into one
