+++ packages/documentation-site/patternfly-docs/content/get-started/release-notes-table.js
@@ -37,27 +37,7 @@ import {
 import FilterIcon from "@patternfly/react-icons/dist/esm/icons/filter-icon";
 import SearchIcon from "@patternfly/react-icons/dist/esm/icons/search-icon";
 
-import { releaseNotes as releaseNotesCard } from "./release-notes-data.js";
-import { releaseNotes as releaseNotesAshley } from "./release-notes-data-ashley.js";
-import { releaseNotes as releaseNotesAustin } from "./release-notes-data-austin.js";
-import { releaseNotes as releaseNotesDallas } from "./release-notes-data-dallas.js";
-import { releaseNotes as releaseNotesEric } from "./release-notes-data-eric.js";
-import { releaseNotes as releaseNotesJenny } from "./release-notes-data-jenny.js";
-import { releaseNotes as releaseNotesKatie } from "./release-notes-data-katie.js";
-import { releaseNotes as releaseNotesNicole } from "./release-notes-data-nicole-portion.js";
-import { releaseNotes as releaseNotesTitani } from "./release-notes-data-titani.js";
-
-const releaseNotes = [
-  ...releaseNotesCard,
-  ...releaseNotesAshley,
-  ...releaseNotesAustin,
-  ...releaseNotesDallas,
-  ...releaseNotesEric,
-  ...releaseNotesJenny,
-  ...releaseNotesKatie,
-  ...releaseNotesNicole,
-  ...releaseNotesTitani
-]
+import { releaseNotes } from "./release-notes-data.js";
 
 export const ReleaseNotesTable = () => {
   // Row expand/collapse
